Log git commit id into sketch file as a comment. Added -Wall. Fixed compiler warnings.

+++ src/rollinghashcpp/characterhash.h
@@ -38,7 +38,7 @@ private:
 template <typename hashvaluetype>
 hashvaluetype maskfnc(int bits) {
     assert(bits>0);
-    assert(bits<=sizeof(hashvaluetype)*8);
+    assert(bits<=(int)sizeof(hashvaluetype)*8);
     hashvaluetype x = static_cast<hashvaluetype>(1) << (bits - 1);
     return x ^ (x - 1);
 }
